Refactor timeline and search into own component
