+++ src/components/timeline_with_search.js
@@ -1,0 +1,86 @@
+import React            from 'react';
+import rehypeReact      from 'rehype-react';
+import Tweet            from '../components/tweet';
+import { Button }       from 'semantic-ui-react';
+import { Container }    from 'semantic-ui-react';
+import { Grid }         from 'semantic-ui-react';
+import { Input }        from 'semantic-ui-react';
+import { Form }         from 'semantic-ui-react';
+import { Timeline }     from 'vertical-timeline-component-for-react';
+import { TimelineItem } from 'vertical-timeline-component-for-react';
+import { useState }     from 'react';
+
+
+const render = new rehypeReact({
+  createElement: React.createElement,
+  components: {
+    'tweet': Tweet
+  }
+}).Compiler;
+
+
+export default function TimelineWithSearch({ data }) {
+  const [ newestFirst, setNewestFirst ]     = useState(true);
+  const [ searchResults, setSearchResults ] = useState([]);
+  const [ searchTerm, setSearchTerm ]       = useState('');
+
+  function handleSearchTermChange(e) {
+    setSearchTerm(e.target.value);
+  }
+
+  React.useEffect(() => {
+    const results = data.filter(({ node: { html, frontmatter } }) => (
+      html.toLowerCase().includes(searchTerm.toLowerCase()) || frontmatter.title.toLowerCase().includes(searchTerm.toLowerCase())
+    ));
+    setSearchResults(results);
+  }, [ data, searchTerm ]);
+
+  return (
+    <Container>
+      <Form>
+        <Grid>
+          <Grid.Column computer={12} tablet={12} mobile={8}>
+            <Form.Field>
+              <label>Search</label>
+              <Input
+                size="mini"
+                icon="search"
+                value={searchTerm}
+                onChange={handleSearchTermChange}
+                autoComplete={false}
+              />
+            </Form.Field>
+          </Grid.Column>
+          <Grid.Column computer={4} tablet={4} mobile={8}>
+            <Form.Field>
+              <label style={{ textAlign: 'left' }}>Sort</label>
+              <Button.Group size="mini">
+                <Button size="mini" color="red" basic={!newestFirst}  onClick={() => setNewestFirst(true)}>Newest</Button>
+                <Button size="mini" color="red" basic={newestFirst} onClick={() => setNewestFirst(false)}>Oldest</Button>
+              </Button.Group>
+            </Form.Field>
+          </Grid.Column>
+        </Grid>
+      </Form>
+      {searchResults.length > 0 && (
+        <Timeline className={newestFirst ? null : 'reverse'} lineColor={'#ddd'} animate={false}>
+          {searchResults.map(({ node }) => {
+            return (
+              <TimelineItem
+                key={node.id}
+                id={node.frontmatter.date}
+                dateText={node.frontmatter.title}
+                dateInnerStyle={{ background: '#B71C1C', color: '#FFF' }}
+                style={{ color: '#B71C1C' }}>
+                {render(node.htmlAst)}
+              </TimelineItem>
+            );
+          })}
+        </Timeline>
+      )}
+        {searchResults.length === 0 && (
+          <p style={{ textAlign: 'center' }}>No Results</p>
+        )}
+    </Container>
+  );
+}
